+++ Invoices/Queries/InvoiceDto.cs
@@ -1,4 +1,4 @@
 ﻿using System;
 namespace Invoices.Queries;
 
-public record InvoiceDto(int Id, DateTime Date, Models.InvoiceStatus Status, decimal Total, decimal Vat, decimal VatRate, decimal? Paid);
+public record InvoiceDto(int Id, DateTime Date, Models.InvoiceStatus Status, decimal Total, decimal Vat, double VatRate, decimal? Paid);
